Module Alsa

* Interface with the alsa drivers. * *

author Samuel Mimram

val get_version : unit -> string

Get the ALSA sound library version in ASCII format.

exception Buffer_xrun

A buffer underrun / overrun occured.

exception Bad_state

PCM is not in the right state.

exception Suspended

A suspend event occurred (stream is suspended and waiting for an application * recovery).

exception IO_error

Input/output error.

exception Device_busy

Device or resource was busy.

exception Invalid_argument

Function was called with an invalid argument.

exception Device_removed

This error can happen when device is physically * removed (for example some hotplug devices like USB * or PCMCIA, CardBus or ExpressCard can be removed on the fly).

exception Interrupted

exception Unknown_error of int

type direction =

| Dir_down

| Dir_eq

| Dir_up

val string_of_error : exn -> string

Get an error message corresponding to an error. * Raise the given exception if it is not known.

val no_stderr_report : unit -> unit

Do not report errors on stderr.

module Pcm : sig ... end

module Sequencer : sig ... end